change to uint256

+++ database/102-up-timescale/20230605110239-lootboxes_update-volume-type.sql
@@ -1,7 +1,7 @@
 -- migrate:up
 
 ALTER TABLE lootbox
-	ALTER COLUMN volume TYPE NUMERIC;
+	ALTER COLUMN volume TYPE uint256;
 
 -- migrate:down
